--- sig/datadog/di/configuration/settings.rbs
@@ -1,10 +1,0 @@
-module Datadog
-  module DI
-    module Configuration
-      module Settings
-        def self.extended: (untyped base) -> untyped
-        def self.add_settings!: (untyped base) -> untyped
-      end
-    end
-  end
-end
